Remove EDH user api methods & tests
